i2c: update img-scb
This includes following changes :-

1. enable fencing for all versions of the ip since 2.2.1
The code to read from the master read fifo, and write to the master
write fifo, checks a bit in an SCB register before every byte to
ensure that the fifo is not full (write fifo) or empty (read fifo).
Due to clock domain crossing inside the SCB block the updated value
of this bit is only visible after 2 cycles.

The scb_wr_rd_fence() function does 2 dummy writes (to the read-only
revision register), and it's called before reading from or writing to the
fifos to ensure that subsequent reads of the fifo status bits do not read
stale values.

The 2 dummy writes are required in all versions of the ip since 2.2.1

2. reset interrupts in img_i2c_soft_reset
Reset interrupt enable register and clear any generated interrupts
to make sure of a clean slate after a soft reset. Not doing so might
leave unhandle line status or generated interrupts which can cause
issues when handling new transfers.

3. improve transaction complete handle
Clear line status and all interrupts when transaction is complete,
as not doing so might leave unserviced interrupts that might be
handled in the context of a new transfer. Soft reset if the the
transfer failed to bring back the i2c block to a reset state.

4. verify support for requested bit rate
The requested bit rate can be outside the range supported by the driver.
The maximum bit rate this driver supports at the moment is 400Khz.

Return -EINVAL if the bit rate is larger than 400khz.

Maximum speed supported by the driver can be increased to 1Mhz by
adding support for "fast plus mode" in the future.

5. fix LOW and HIGH period values for the SCL clock
After determining the minimum value for the High period (TCKH) the
remainder of the internal clock pulses is set as the Low period (TCKL).
This causes the i2c clock duty cycle to be much less than 50%.

The fix suggested here, start with TCKH and TCKL at 50% of the internal
clock pulses and adjusts the TCKH and TCKL values from there if the
minimum value for TCKL is not met. This will make sure the i2c clock
duty cycle is at 50% or close 50% whenever possible.

6. support I2C_M_IGNORE_NAK
This commit adds support for the I2C_M_IGNORE_NAK protocol
modification.

Such behaviour can only be implemented in atomic mode. So, if a
transaction contains a message with such flag the drivers
switches to atomic mode. The implementation consists simply in
treating NAKs as ACKs.

7. i2c: img-scb: call img_i2c_soft_reset after a time out
Call img_i2c_soft_reset after a transfer has timed out to bring
back the i2c block to a clean stale.

Not doing so could cause all pending segments on a list of transfers
to fail.

8. calls to img_i2c_transaction_halt relocated to support and control repeated starts on IP version 3.3.
In version 3.3 of the IP when transaction halt is set, an interrupt
will be generated after each byte of a transfer instead of after
every transfer but before the stop bit.
Due to this behaviour we have to be careful that every time we
release the transaction halt we have to re-enable it straight away
so that we only process a single byte, not doing so will result in
all remaining bytes been processed and a stop bit being issued,
which will prevent us having a repeated start.

This change will have no effect on earlier versions of the IP.

9. remove start bit detected status after handling
Remove start bit detected status after it is handled,
doing so will prevent this condition being hit for
every interrupt on a particular transfer.

10. mark transaction as complete when all data is read
We can mark the transfer as complete without waiting for the stop
bit. This is important when handling repeated start transfers as
we have to start the next transfer without the stop bit being issued.

11. mark transaction as complete when no more data to write
We can mark the transfer as complete without waiting for the stop
bit. This is important when handling repeated start transfers as
we have to start the next transfer without the stop bit being issued.

This doesn't effect the older versions of the IP.

12. remove fifo EMPTYING interrupts handle
This interrupt could have been useful for repeated start transfers
as the current transfer could be marked as complete while it's
processing the final byte of the transfer.
But having to use the transaction halt interrupt to safely control
repeated start transfers, means handling of the fifo EMPTYING
interrupts is no longer necessary.

Handling this interrupt along with Transaction Halt interrupt can
cause erratic behaviour.

13. add handle for stop detected interrupt
Triggered when a Stop bit is detected on the bus. No special
requirement on this interrupt but is just an indicator of end
of current transfer.
Stop bit indicates the end of a transfer and if the transfer
has finished before all data is written to the fifo or with more
data expected return error with transfer complete signal.

14. add handle for Master halt interrupt
Master halt is issued after each byte of a transaction is processed,

Master halt will stall the bus by holding the SCK line low until halt
bit in scb_general_control is cleared.

This interrupt can be used to control repeated starts where we can
catch the end of the transfer without the stop bit being issued.

Change-Id: I0cf1209c5a512f05026a4de503c1a36299737517
Signed-off-by: Sifan Naeem <sifan.naeem@imgtec.com>