pistachio: bt: fixes
This change fixes :-

1. plat. driver section mismatch
Since probe() routine is in __init section and we use
platform_driver_probe(), there is no need for .probe field in the
img_bt_driver struct. This gets rid of the linker warning.

BUG=none
TEST='modprobe img-bt'

2. pistachio: bt: fix a reference to an old header
"img-transport.h" has moved to a new location and this #include
hasn't been updated.

BUG=none
TEST=kernel recompilation

Change-Id: I6a0758b8bf3d2b7da57d06a832ccb52926b2860c
Signed-off-by: Bartosz Flis <bartosz.flis@imgtec.com>