dma: update mdc
This include following changes :-

1. dma: mdc: add suspend resume support.
Add support for mdc dev pm ops with suspend-resume,
mdc clocks can be gated when device suspends.

2. dma: img: Revert channel enable/disable
Revert a previous patch that changed the ordering of SoC level
enable/disable of channels to match upstream.

Both orderings are valid.

3. dma: mdc: Correct event check comparison
The event count check should be >= to cause a return
of 0 for no new events

Presently it returns the maximum number of events
when there are no new events

Change-Id: I09afba598bfb45bdfea7b64b78ff313e06e506ed
Signed-off-by: Damien.Horsley <Damien.Horsley@imgtec.com>
Signed-off-by: Govindraj Raja <Govindraj.Raja@imgtec.com>